refactor: modity dto
