+++ src/auth/dto/payload.ts
@@ -1,4 +1,4 @@
-import { Gender } from "src/news/common/Gender";
+import { Gender } from "src/news/common/gender.enum";
 import { Social } from "../common/Social";
 
 export class Payload {
